Exclude BDD tests from sonarcloud run
The BDD tests included are a working MVP. They do require for the
docker-compose setup to be up and running though. They also take quite
some time to run (even for the three tests currently present). SO should
this be in the responsibility of the developer to run these? Or should
this be part of the SonarCloud/Github pipeline as well?