Refactor and introduce BDD tests MVP (#30)
* Refactor and introduce BDD tests MVP

* Condiitonally replace localhost with dockerhost to (hopefully) enable

* test passing on github actions
* NOTE they do work on local

* Try again in a different way getting the local ip

* Try timeout

* Try with retries for request session

* Correct construct_zkp attr type

* add another test for empty input

* Exclude BDD tests from sonarcloud run

The BDD tests included are a working MVP. They do require for the
docker-compose setup to be up and running though. They also take quite
some time to run (even for the three tests currently present). SO should
this be in the responsibility of the developer to run these? Or should
this be part of the SonarCloud/Github pipeline as well?